Revert "enable the repair controller in the service controller"
There must no need to sync new services, that is exactly what the
informer does, if we are worried about it, we just enable the
resync period in the informer.

However, there can still be the case that OVN has stale data,
but this may be overkill for dealing with that.

This reverts commit 9029348caa74050adf7dcda497d441448087ed11.